separate wanco.c
