+++ lib/Image.h
@@ -1,0 +1,6 @@
+// Image
+typedef struct Image{
+    int w;
+    int h;
+    char buf[IMG_MAX_W*IMG_MAX_H];
+} Image;
