+++ wanco_res.h
@@ -1,0 +1,107 @@
+// Game resouces
+Image resBtnWanco0 = {
+    .w   = 7,
+    .h   = 3,
+    .buf = "-------"
+           " wanco "
+           "-------"
+};
+Image resBtnWanco1 = {
+    .w   = 7,
+    .h   = 3,
+    .buf = "======="
+           "|wanco|"
+           "======="
+};
+
+Image resBtnNoppo0 = {
+    .w   = 7,
+    .h   = 3,
+    .buf = "-------"
+           " noppo "
+           "-------"
+};
+Image resBtnNoppo1 = {
+    .w   = 7,
+    .h   = 3,
+    .buf = "======="
+           "|noppo|"
+           "======="
+};
+
+Image resGhost = {
+    .w   = 4,
+    .h   = 3,
+    .buf = " __ "
+           "('')"
+           " || "
+};
+
+Image resWanco = {
+    .w   = 8,
+    .h   = 4,
+    .buf = "\0/----\\\0"
+           "| O  O |"
+           "|  ==  |"
+           "\0\\----/\0"
+};
+
+Image resNoppo = {
+    .w   = 6,
+    .h   = 10,
+    .buf = "\0/--\\\0"
+           "|O  O|"
+           "|    |"
+           "| == |"
+           "|    |"
+           "|    |"
+           "|    |"
+           "|    |"
+           "|    |"
+           "\0\\--/\0"
+};
+
+Image resHouse = {
+    .w   = 5,
+    .h   = 8,
+    .buf = "\0\0_\0\0"
+           "\0/ \\\0"
+           "/ _ \\"
+           "| + |"
+           "| = |"
+           "| _ |"
+           "|! !|"
+           "||_||"
+};
+
+Image resHouseX = {
+    .w   = 5,
+    .h   = 8,
+    .buf = "\0\0_\0\0"
+           "\0/ \\\0"
+           "/ _ \\"
+           "| x |"
+           "| = |"
+           "| _ |"
+           "|! !|"
+           "||_||"
+};
+
+Image resMeter = {
+    .w   = 12,
+    .h   = 1,
+    .buf = "[##########]"
+};
+
+Image resGround = {
+    .w   = 96,
+    .h   = 8,
+    .buf = "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^"
+           "                                                                                                "
+           "                                                                                                "
+           "                                                                                                "
+           "                                                                                                "
+           "                                                                                                "
+           "                                                                                                "
+           "                                                                                                "
+};
